% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFounders.R
\name{makeFounders}
\alias{makeFounders}
\title{makeFounders(): makes a matrix of individuals to act as a founding population.}
\usage{
makeFounders(pop = 1000, osr = c(0.5, 0.5), stocks = c(0.3, 0.3,
  0.4), maxAge = 20, survCurv = 0.7^(1:maxAge)/sum(0.7^(1:maxAge)))
}
\arguments{
\item{pop}{The size of the founder population.}

\item{osr}{A numeric vector describing the sex ratio, c([male], [female]). Must sum to 1.}

\item{stocks}{A numeric vector describing the probability that an individual
is in each stock. Must sum to 1.}

\item{maxAge}{Numeric. The max age to which an animal may survive.}

\item{survCurv}{Numeric vector. Describes the probability within the founder cohort of belonging
to each age-class for age=-classes 1:maxAge. Must sum to 1. Cannot be blank.}
}
\description{
Returns a pop-by-8 character matrix, with each row being an individual in the
founder population.
[,1] is a unique (uuid) identifier for each animal.
[,2] is sex, values "M" or "F".
[,3] is "founder" for all animals.
[,4] is "founder" for all animals.
[,5] is "-1" for all animals.
[,6] is NA for all animals.
[,7] is the stock membership for each animal.
[,8] is the age of each animal (in 'breeding seasons') at year 0.
}
