% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_dead.R
\name{remove_dead}
\alias{remove_dead}
\title{remove_dead(): removes the dead from an individual-data matrix.}
\usage{
remove_dead(indiv = mort())
}
\arguments{
\item{indiv}{A matrix of individuals, as from mort().}
}
\description{
For larger simulations, the matrix 'indiv' may grow very large and slow down the simulation.
In these cases, run-times may be improved by periodically moving dead individuals into an
archive that is read and written less frequently than 'indiv'. See also archive_dead().
}
